+++ benchmark/query_benchmark.cc
@@ -49,7 +49,7 @@ class IndexBenchmark {
     void CreateQuery(PhBox<DIM>& query);
 
     const TestGenerator data_type_;
-    const int num_entities_;
+    const size_t num_entities_;
     const double avg_query_result_size_;
 
     constexpr int query_endge_length() {
@@ -94,8 +94,8 @@ template <dimension_t DIM>
 void IndexBenchmark<DIM>::SetupWorld(benchmark::State& state) {
     logging::info("Setting up world with {} entities and {} dimensions.", num_entities_, DIM);
     CreatePointData<DIM>(points_, data_type_, num_entities_, 0, GLOBAL_MAX);
-    for (int i = 0; i < num_entities_; ++i) {
-        tree_.emplace(points_[i], i);
+    for (size_t i = 0; i < num_entities_; ++i) {
+        tree_.emplace(points_[i], (int)i);
     }
 
     state.counters["total_result_count"] = benchmark::Counter(0);
